requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
reportlab>=4.1.0
Pillow>=10.0.0
